requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
google-genai
sentence-transformers
python-dotenv